DE1518 Display Order Information From Service
DE1518 Refactored Out Generally Useful Methods
DE1518 Add Order Detail Request Model
DE1518 Add XSD Name to Validation Log Message
DE1518 Add Detail Request Related Config
DE1518 Add XML Mapper Model
DE1518 Add Data Extraction Mappings To Config
DE1518 Modify View Template For Status History
Set necessary fields per payment method.
DE1518 Update Payment CC types config
DE1518 Set Shipping Method Correctly
DE1518 Add Extraction Method for Summing Data
DE1518 Display Tax and Totals
DE1518 Modify Items To Display Gift Messages
DE1518 Don't Load Order If Request Fails
DE1518 Cache Order Search Results In Registry
Add sales order shipment adapter to replace
magento data with OMS order detail data.
Override sales order print block, print shipment and shipping tracking popup block.
DE1518 Setup Item Options to Be Displayed
DE1518 Make Shipment/Tracking Use Order Detail
Make sure tracking is popup page is using order detail
data. Make sure links that required data from actual
Magento model get order id instead.

+++ src/app/code/community/EbayEnterprise/Eb2cOrder/Model/Eav/Entity/Increment/Order.php
@@ -40,7 +40,7 @@ class EbayEnterprise_Eb2cOrder_Model_Eav_Entity_Increment_Order
 	public function getNextId()
 	{
 		// remove any order prefixes from the last increment id
-		$last = (int) Mage::helper('eb2corder')->removeOrderIncrementPrefix($this->getLastId());
+		$last = Mage::helper('eb2corder')->removeOrderIncrementPrefix($this->getLastId());
 		$next = $last + 1;
 		return $this->format($next);
 	}
